+++ tests/testthat/test-cnd-abort.R
@@ -268,3 +268,18 @@ test_that("report-specific backtrace-on-error option has precedence", {
     expect_equal(peek_backtrace_on_error(), "full")
   )
 })
+
+test_that("abort() displays call in error prefix", {
+  skip_if_not_installed("rlang", "0.4.11.9001")
+
+  expect_snapshot(
+    run("rlang::abort('foo', call = quote(bar(baz)))")
+  )
+
+  # errorCondition()
+  skip_if_not_installed("base", "3.6.0")
+
+  expect_snapshot(
+    run("rlang::cnd_signal(errorCondition('foo', call = quote(bar(baz))))")
+  )
+})